Added Filter Functionality, Changed intensity to int

+++ src/CS3500/model/image/IImage.java
@@ -22,8 +22,7 @@ public interface IImage {
    * @return a {@link IMatrix} consisting of the intensity values of each pixel's specified channel.
    * @throws IllegalArgumentException if the supplied channel is {@code null}.
    */
-  // IMatrix<Integer> extractChannel(EChannelType channel)
-  IMatrix<Double> extractChannel(EChannelType channel)
+  IMatrix<Integer> extractChannel(EChannelType channel)
       throws IllegalArgumentException;
 
   /**
